Extracted the routes from JawsController in separate traits
